convert NewConnectorModal to component
